refactor: extract active workout UI into dedicated component and show loading state
