+++ src/components/ui/loading-state.tsx
@@ -1,0 +1,31 @@
+import { Text } from '@/src/components/ui/text';
+import { cn } from '@/src/lib/utils/cn';
+import { colors } from '@/src/theme/tokens';
+import { ActivityIndicator, View } from 'react-native';
+
+type LoadingStateProps = {
+  label?: string;
+  className?: string;
+  size?: 'small' | 'large';
+};
+
+export function LoadingState({
+  label = 'Loading...',
+  className,
+  size = 'large'
+}: LoadingStateProps) {
+  return (
+    <View
+      accessibilityLabel={label}
+      accessibilityRole="progressbar"
+      className={cn('flex-1 items-center justify-center px-6 py-10', className)}
+    >
+      <ActivityIndicator color={colors.primary} size={size} />
+      {label ? (
+        <Text variant="small" tone="muted" className="mt-3 text-center">
+          {label}
+        </Text>
+      ) : null}
+    </View>
+  );
+}
